wip: Get a very rough working version that sets store with fetched categories/data.
For what ever reason, I found this to be very difficult.
Reference the advanced redux doc examples for further
help and to clean this up.

Basic idea was to fetch the data in the componenet for
the categories. Then load that into the redux store
as it will be used across multiple componenets (at
least that's the plan for now).