Setup Link component for post details.
Add Link components back in to let a user
click a post title and go to the detail
page for that post.

This is a bit different setup with redux
compared to previous uses of BrowserRouter.
I had to Use Router with a Route for the main
app in my index.js then setup Routes like
before in my app.js (main component).