wip: Break App component into smaller ones.
App component was becoming large and handled
too many different responsibilities. I've broken
it down to handle the routes and it's original
content is now 2 separate components for now.

Also had to remove unused setup from App component
which was causing Router issues and would not
let the Post component load as a separate page.