A bit of polish.
Update UI to be a bit more presentable
and easier to use.

Switch links and buttons out to icons or
better styled buttons.

Add back links with arrow icons to help
users navigate.

Update redux store when comments and posts
are deleted to trigger rendering and update
the UI with the removed item.

Cleanup some code formatting (this udacity
built in editor is not bad but could use
some updates).

Add FontAwesome support.

Update how forms update and set state to avoid
warnings not using controlled components and
fix editing a post to allow the title to be
pre-loaded into the form.